+++ frontend/svelte/src/routes/ProjectDetail.svelte
@@ -12,13 +12,15 @@
   } from "../lib/stores/layout.store";
   import MainContentWrapper from "../lib/components/ui/MainContentWrapper.svelte";
   import {
-    loadSnsFullProject,
+    loadSnsSummary,
+    loadSnsSwapStateStore,
     routePathRootCanisterId,
   } from "../lib/services/sns.services";
   import { isRoutePath } from "../lib/utils/app-path.utils";
   import {
     type SnsFullProject,
     snsFullProjectStore,
+    snsSummariesStore,
   } from "../lib/stores/projects.store";
   import { getSnsProjectById } from "../lib/utils/sns.utils";
   import Spinner from "../lib/components/ui/Spinner.svelte";
@@ -49,7 +51,7 @@
     }
     rootCanisterIdString = rootCanisterIdMaybe;
 
-    await loadSnsFullProject(rootCanisterIdString);
+    await loadSnsSummary(rootCanisterIdString);
   });
 
   onDestroy(unsubscribe);
@@ -62,6 +64,11 @@
   layoutBackStore.set(goBack);
 
   $: layoutTitleStore.set(fullProject?.summary.name ?? "");
+
+  // TODO(L2-838): if error redirect to launchpad and display error there
+
+  // TODO: do we want such subscribe in the component?
+  $: $snsSummariesStore, (() => loadSnsSwapStateStore(rootCanisterIdString))();
 </script>
 
 {#if fullProject === undefined}